Revert "Created the initial code for database. Restructured the files."
