--- .gitignore
@@ -1,4 +1,0 @@
-website/__pycache__
-virtualenv
-alembic.ini
-migrations
