--- requirements.txt
@@ -1,13 +1,0 @@
-alembic==1.7.7
-click==8.0.4
-colorama==0.4.4
-Flask==2.0.3
-Flask-SQLAlchemy==2.5.1
-greenlet==1.1.2
-itsdangerous==2.1.1
-Jinja2==3.0.3
-Mako==1.2.0
-MarkupSafe==2.1.1
-psycopg2==2.9.3
-SQLAlchemy==1.4.32
-Werkzeug==2.0.3
