+++ templates/index.html
@@ -6,8 +6,8 @@
 {% block body %}
 <div class="content" align="middle">
     <img src="{{ url_for('static', filename='lock.jpg')}}" align="middle" />
-    <p>AZORIAN</p>
-    <a href="{{ url_for('views.new_form') }}">Request Authority to Operate</a>
+    <p>AZONIAN</p>
+    <a href="{{ url_for('new_form') }}">Request Authority to Operate</a>
 
 </div>
 {% endblock %}
