+++ templates/new_form.html
@@ -145,7 +145,7 @@
             Please ensure all information on this form is accurate to the best of your knowledge.
         </p>
         <div class="btn-group">
-          <a class="btn-item" href="https://www.epa.gov/developers/authorization-operate-overview" target="_blank">What is Authority to Operate?</a>
+          <a class="btn-item" href="https://www.epa.gov/developers/authorization-operate-overview">What is Authority to Operate?</a>
         </div>
       </div>
       <form action="/">
